backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
python-dotenv
django-cors-headers
ruff
black
pytest
pytest-django
openai